8.12 StrI(value as Integer, radix as Integer) as String
Converts the integer value into a string representation using the given radix.

If radix is not 2 .. 36 then an empty string is returned.

Note that the returned string does not include a base prefix and uses lowercase letters to represent those digits in bases greater than 10.

Example:

    print StrI(255, 16) '= "ff"

    print StrI(9, 2) '= "1001"

This function is not available in the 3.1 firmware version.